Split norent/routes.tsx from norent/site.tsx.
